move naming logic into etl module
